Rename crystalfontz.dbus.map to crystalfontz.dbus.domain
